requirements.txt
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
imageio==2.34.0
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.3
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.0.3
pillow==10.2.0
ptflops==0.7.2.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-fid==0.1.1
pytz==2024.1
PyWavelets==1.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
tifffile==2023.7.10
torch==2.2.1
torchsummary==1.5.1
torchvision==0.17.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
zipp==3.17.0
